Preparing Cross-tab communication for Iframe contents.
The stage is set for implementation of communication across tabs to return iframe elements as a copy to the original instance of the program. This has to be done using the tab IDs on a page, or by sending information to the server for use. The future logic is going to have to follow this general outline:

The "new" tab returns the elements as well as the URL associated with the iframe. When the "original" tab receives this information, it is going to have to check whether the URL being returned from the "new" tab matches the URL of the iframe on its page. If it does, it is a match and we know we are receiving the right information.

This, unfortunately, will not work well on pages like ObservableHQ, which don't host the D3 scripts within the iframe itself.