Updated the parser to dynamically work with types in config file.
The JSON config file "SupportedTypes.json" is currently a barebones list of supported types, their name, and the functions commonly associated with them. This list can be updated by adding type additions to the JSON file itself. The file will become more complex as time goes on so the characteristics of certain visualization types can be more accurately catalogued.